fixed apiUrl construction, separated interfaces, fixed environment.example.ts
