+++ src/app/pages/home/home.component.html
@@ -5,7 +5,7 @@
     <div class="refresh-search-wrapper">
       <button
         class="refresh"
-        (click)="fetchAndDisplayData()"
+        (click)="fetchAndDisplayData('/JobPost')"
         mat-stroked-button
       >
         Refresh
