+++ src/environment/environment.example.ts
@@ -1,3 +1,4 @@
 export const environment = {
-  bearerToken: 'xxxx',
+  bearerToken: ' Bearer xxxx',
+  baseUrl: 'https://stage.api.recruitment.indigo.si',
 };
